Added parent component. Moved transform components.
